fix: 좋아요 개수 반환 long 타입으로 수정

+++ backend/src/main/java/kr/touroot/travelogue/repository/query/TravelogueLikeQueryRepository.java
@@ -2,5 +2,5 @@ package kr.touroot.travelogue.repository.query;
 
 public interface TravelogueLikeQueryRepository {
 
-    int countTravelougeLikeByRank(int rank);
+    long countTravelougeLikeByRank(int rank);
 }
